% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_sem_power_graph.R
\name{pls_sem_power_graph}
\alias{pls_sem_power_graph}
\title{Plot graph for PLS-SEM power analysis or sensitivity}
\usage{
pls_sem_power_graph(
  method = "a priori",
  MDES = NULL,
  N = NULL,
  alpha = 0.05,
  tails = 1,
  theme = "abyss"
)
}
\arguments{
\item{method}{"a priori" (compute sample size) or "sensitivity" (compute MDES)}

\item{MDES}{Minimum Detectable Effect Size (required if method = "a priori")}

\item{N}{Sample size (required if method = "sensitivity")}

\item{alpha}{Significance level. Must be one of 0.01, 0.05, 0.10}

\item{tails}{Number of tails: 1 (default) or 2}

\item{theme}{Plot theme: "abyss" (default) or "min"}
}
\value{
A ggplot object
}
\description{
Plot graph for PLS-SEM power analysis or sensitivity
}
