% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_sem_power.R
\name{pls_sem_power}
\alias{pls_sem_power}
\title{Power or Sensitivity Calculation for PLS-SEM}
\usage{
pls_sem_power(
  method = c("sensitivity", "a priori"),
  N = NULL,
  MDES = NULL,
  alpha = 0.05
)
}
\arguments{
\item{method}{Either "a priori" or "sensitivity"}

\item{N}{Sample size (required for sensitivity)}

\item{MDES}{Minimum detectable effect size (required for a priori)}

\item{alpha}{Significance level: one of 0.01, 0.05, or 0.10}
}
\value{
Printed result
}
\description{
Power or Sensitivity Calculation for PLS-SEM
}
